requirements.txt
streamlit
python-jose
passlib[bcrypt]
sqlite3
